requirements.txt
fastapi
uvicorn
openai
tiktoken
pymongo
python-multipart
pydantic
python-dotenv
